Check if file exists in analysis location before uploading

+++ steerPIDAnalysis.sh
@@ -10,7 +10,14 @@ jetString=$7
 modeString=$8
 prePIDMode=$9
 
+analysisFileName=$(basename $analysisFile)
+
 ssh $remoteHost mkdir -p $remoteBasePath/$analysisDir
-scp runPIDAnalysis.sh $remoteHost://$remoteBasePath
-scp $analysisFile $remoteHost://$remoteBasePath/$analysisDir
+if ! ssh -q $remoteHost "test -e $remoteBasePath/$analysisDir/$analysisFileName";then
+    echo "Analysis file does not exist, uploading..."
+    scp $analysisFile $remoteHost://$remoteBasePath/$analysisDir
+else
+    echo "Analysis file $analysisFile already exists in remote, using existing file"
+fi
+
 ssh $remoteHost $remoteBasePath/runPIDAnalysis.sh $analysisDir $(basename $analysisFile) $identifier \"$centString\" \"$jetString\" \"$modeString\" \"$prePIDMode\"
